[fixup] switch to service provider
