+++ integration-tests/src/test/resources/META-INF/services/com.datastax.oss.driver.api.mapper.result.MapperResultProducerRegistry
@@ -1,0 +1,1 @@
+com.datastax.oss.driver.mapper.GuavaFutureProducerRegistry
