+++ mapper-processor/src/main/java/com/datastax/oss/driver/internal/mapper/processor/mapper/MapperBuilderGenerator.java
@@ -76,8 +76,7 @@ public class MapperBuilderGenerator extends SingleFileCodeGenerator {
                     .returns(ClassName.get(interfaceElement))
                     .addStatement(
                         "$1T context = new $1T(session, defaultKeyspaceId, "
-                            + "defaultExecutionProfileName, defaultExecutionProfile, "
-                            + "customState, resultProducers)",
+                            + "defaultExecutionProfileName, defaultExecutionProfile, customState)",
                         DefaultMapperContext.class)
                     .addStatement(
                         "return new $T(context)",
